requirements.txt
streamlit
snowflake-connector-python
pandas
fastapi
uvicorn
numpy
requests